[BE] feat: 게시글 조회, 정렬, 마이페이지 조회 로직 추가

+++ backend-spring/src/main/java/com/ssafy/em/posts/dto/PostDetailDto.java
@@ -12,7 +12,7 @@ public record PostDetailDto(
         String imageUrl,
         String content,
         Point location,
-        Map<String, Integer> emotionCountList,
+        Map<String, Long> emotionCountList,
         LocalDateTime createdAt
 ) {
 }
